Enum search_graph::PathItem

pub enum PathItem<'a, T, S, A> where T: 'a + Hash + Eq + Clone, S: 'a, A: 'a {
    Item(Edge<'a, T, S, A>),
    Head(Target<Node<'a, T, S, A>, Edge<'a, T, S, A>>),
}

Sum type for path elements. All elements except the head are represented with the PathItem::Item variant.

Variants

Item

Non-head item, a (vertex, edge) pair.

Head

The path head, which may resolve to a vertex or an unexpanded edge.