Module search_graph::mutators

Support for navigation of a graph that allows modifications to graph topology and full read-write access to graph data.

The data structures in this module require a read-write borrow of an underlying graph. As a result, only one handle may be active at any given time.

Structs

MutChildList
A traversible list of a vertex's outgoing edges.

MutEdge
Mutable handle to a graph edge ("edge handle") when edge expansion state is unknown.

MutNode
Mutable handle to a graph vertex ("node handle").

MutParentList
A traversible list of a vertex's incoming edges.