Struct search_graph::mutators::MutEdge

pub struct MutEdge<'a, T: Hash + Eq + Clone + 'a, S: 'a, A: 'a> { /* fields omitted */ }

Mutable handle to a graph edge ("edge handle").

This zipper-like type enables traversal of a graph along the edge's source and target vertices.

It enables local graph mutation, whether via mutation of edge data or mutation of graph topology (adding vertices).

Methods

impl<'a, T: Hash + Eq + Clone + 'a, S: 'a, A: 'a> MutEdge<'a, T, S, A>

pub fn get_id(&self) -> usize

Returns an immutable ID that is guaranteed to identify this vertex uniquely within its graph. This ID may change when the graph is mutated.

pub fn get_data(&self) -> &A

Returns the data at this edge.

pub fn get_data_mut(&mut self) -> &mut A

Returns the data at this edge, mutably.

pub fn get_target<'s>(&'s self) -> Node<'s, T, S, A>

Returns the target of this edge. Returns a node handle, whose lifetime is limited to a local borrow of self.

pub fn get_target_mut<'s>(&'s mut self) -> MutNode<'s, T, S, A>

Returns the target of this edge. Returns a mutable node handle will be available, whose lifetime is limited to a local borrow of self.

pub fn to_target(self) -> MutNode<'a, T, S, A>

Returns the target of this edge. Consumes self and returns a mutable node handle whose lifetime will be the same as that of self.

pub fn get_source<'s>(&'s self) -> Node<'s, T, S, A>

Returns a node handle for the source of this edge. Its lifetime will be limited to a local borrow of self.

pub fn get_source_mut<'s>(&'s mut self) -> MutNode<'s, T, S, A>

Returns a mutable node handle for the source of this edge. Its lifetime will be limited to a local borrow of self.

pub fn to_source(self) -> MutNode<'a, T, S, A>

Returns a mutable node handle for the source of this edge. self is consumed, and the return value's lifetime will be equal to that of self.

pub fn to_edge(self) -> Edge<'a, T, S, A>

Returns a non-mutating edge obtained by converting this edge. self is consumed, and the return value's lifetime will be the same as that of self. The source graph is still considered to have a mutable borrow in play, but the resulting edge can be cloned freely.