Struct search_graph::mutators::MutParentList

pub struct MutParentList<'a, T, S, A> where T: Hash + Eq + Clone + 'a, S: 'a, A: 'a {
    // some fields omitted
}

A traversible list of a vertex's incoming edges.

Methods

impl<'a, T, S, A> MutParentList<'a, T, S, A> where T: Hash + Eq + Clone + 'a, S: 'a, A: 'a

fn len(&self) -> usize

Returns the number of incoming edges.

fn is_empty(&self) -> bool

Returns true iff there are no incoming edges.

fn get_target_node<'s>(&'s self) -> Node<'s, T, S, A>

Returns a node handle for the vertex these edges originate terminate on. Its lifetime will be limited to a local borrow of self.

fn get_target_node_mut<'s>(&'s mut self) -> MutNode<'s, T, S, A>

Returns a mutable node handle for the vertex these edges terminate on. Its lifetime will be limited to a local borrow of self.

fn to_target_node(self) -> MutNode<'a, T, S, A>

Returns a mutable node handle for the vertex these edges terminate on. self is consumed, and the return value's lifetime will be the same as that of self.

fn get_edge<'s>(&'s self, i: usize) -> Edge<'s, T, S, A>

Returns a handle to the ith edge. Its lifetime will be limited to a local borrow of self.

fn get_edge_mut<'s>(&'s mut self, i: usize) -> MutEdge<'s, T, S, A>

Returns a mutable handle to the ith edge. Its lifetime will be limited to a local borrow of self.

fn to_edge(self, i: usize) -> MutEdge<'a, T, S, A>

Returns a mutable handle to the ith edge. self is consumed, and the return value's lifetime will be the same as that of self.

fn iter<'s>(&'s self) -> ParentListIter<'s, T, S, A>

Returns an iterator over parent edges.