Module search_graph::nav

Support for navigation of a graph without allowing modifications to graph topology.

The data structures in this module require a read-only borrow of an underlying graph. It is safe to navigate through a graph with such multiple structures pointing into it. Graph data that may be updated safely through read-only references (such as atomic types and std::cell::RefCell) may be modified through these structures.

Structs

ChildList
A traversible list of a vertex's outgoing edges.

ChildListIter
Iterator over a vertex's child edges.

Edge
Immutable handle to a graph edge ("edge handle").

Node
Immutable handle to a graph vertex ("node handle").

ParentList
A traversible list of a vertex's incoming edges.

ParentListIter
Iterator over a vertex's parent edges.