Struct search_graph::nav::Edge

pub struct Edge<'a, T, S, A> where T: Hash + Eq + Clone + 'a, S: 'a, A: 'a {
    // some fields omitted
}

Immutable handle to a graph edge ("edge handle").

This zipper-like type enables traversal of a graph along the edge's source and target vertices.

Methods

impl<'a, T, S, A> Edge<'a, T, S, A> where T: Hash + Eq + Clone + 'a, S: 'a, A: 'a

fn get_id(&self) -> usize

Returns an immutable ID that is guaranteed to identify this edge uniquely within its graph. This ID may change when the graph is mutated.

fn get_data(&self) -> &'a A

Returns the data at this edge.

fn get_source(&self) -> Node<'a, T, S, A>

Returns a node handle for this edge's source vertex.

fn get_target(&self) -> Target<Node<'a, T, S, A>, ()>

Returns the target of this edge. If the edge is unexpanded, no data will be available. If it is expanded, a node handle will be available.