Struct search_graph::nav::Node

pub struct Node<'a, T, S, A> where T: Hash + Eq + Clone + 'a, S: 'a, A: 'a {
    // some fields omitted
}

Immutable handle to a graph vertex ("node handle").

This zipper-like type enables traversal of a graph along the vertex's incoming and outgoing edges.

Methods

impl<'a, T, S, A> Node<'a, T, S, A> where T: Hash + Eq + Clone + 'a, S: 'a, A: 'a

fn get_id(&self) -> usize

Returns an immutable ID that is guaranteed to identify this vertex uniquely within its graph. This ID may change when the graph is mutated.

fn get_data(&self) -> &'a S

Returns the data at this vertex.

fn is_leaf(&self) -> bool

Returns true iff this vertex has no outgoing edges (regardless of whether they are expanded).

fn is_root(&self) -> bool

Returns true iff this vertex has no incoming edges.

fn get_child_list(&self) -> ChildList<'a, T, S, A>

Returns a traversible list of outgoing edges.

fn get_parent_list(&self) -> ParentList<'a, T, S, A>

Returns a traversible list of incoming edges.