Enum search_graph::search::SearchError

pub enum SearchError<E> where E: Error {
    Unexpanded,
    ChildBounds {
        requested_index: usize,
        child_count: usize,
    },
    ParentBounds {
        requested_index: usize,
        parent_count: usize,
    },
    SelectionError(E),
}

Errors that may arise during search.

Variants

Unexpanded

A traversal operation could not be performed because the path head is unexpanded.

ChildBounds

A search operation selected a child index that was out of bounds.

Fields

requested_index

The index of the child that was requested.

child_count

The actual number of chidren (which requested_index exceeds).

ParentBounds

A search operation selected a parent index that was out of bounds.

Fields

requested_index

The index of the parent that was requested.

parent_count

The actual number of parents (which requested_index exceeds).

SelectionError

A search operation encountered an error.