Struct search_graph::EdgeExpander

pub struct EdgeExpander<'a, T, S, A> where T: Hash + Eq + Clone + 'a, S: 'a, A: 'a {
    // some fields omitted
}

Modifies graph topology by connecting an unexpanded edge to its target vertex.

An unexpanded edge is one with known source but unknown target. Expanding an edge may connect it to an existing vertex (in which case O(|V|) cycle detection is done) or to a new one.

Methods

impl<'a, T, S, A> EdgeExpander<'a, T, S, A> where T: Hash + Eq + Clone + 'a, S: 'a, A: 'a

fn get_edge<'s>(&'s self) -> Edge<'s, T, S, A>

Returns an edge handle for the edge that this expander would expand. Its lifetime will be limited to a local borrow of self.

fn get_edge_mut<'s>(&'s mut self) -> MutEdge<'s, T, S, A>

Returns a mutable edge handle for the edge that this expander would expand. Its lifetime will be limited to a local borrow of self.

fn to_edge(self) -> MutEdge<'a, T, S, A>

Returns a mutable edge handle for the edge that this expander would expand. self is consumed, and the return value's lifetime will be the same as that of self.

fn expand_to_edge<G>(self, state: T, g: G) -> MutEdge<'a, T, S, A> where G: FnOnce() -> S

Expands this expander's edge, by connecting to the vertex associated with the game state state.

If state does not correspond to an extant vertex, a new vertex will be added for state, initialized with the data produced by g.

Returns an edge handle for the newly expanded edge.

fn expand_to_target<G>(self, state: T, g: G) -> MutNode<'a, T, S, A> where G: FnOnce() -> S

Expands this expander's edge, by connecting to the vertex associated with the game state state.

If state does not correspond to an extant vertex, a new vertex will be added for state, initialized with the data produced by g. A parent edge pointing back to the vertex that this edge originates from will also be added, with initial data the value returned by h.

Returns a node handle for the newly expanded edge's target.