Struct search_graph::Graph

pub struct Graph<T: Hash + Eq + Clone, S, A> { /* fields omitted */ }

A directed graph over a space of discrete, enumerated states.

In typical usage, vertices in the graph will correspond to game states, and edges will correspond to game actions. Vertices have an associated state (e.g., the configuration of pieces on a chessboard) and data (e.g., upper and lower bounds of the score of the outcomes that can be reached from that state), while edges simply have an associated data item (which may include additional statistics statistics or game-specific data about the action that the edge represents).

A Graph is parameterized by three types:

T: The type of game states. It is required to derive Hash and Eq to so that it may be stored in a hashtable, where game states are looked up to support de-duplication of game states. It is required to derive Clone to accommodate a limitation of the HashMap interface.

S: The vertex data type.

A: The edge data type.

Vertices are addressable by content. Cursors into the graph may be obtained with find_node or find_node_mut.

Content can be added to a Graph directly with the [add_node](struct.Graph.html#method.add_node and add_edge methods. It may also be added through the interfaces provided by the mutators/index.html and view/index.html modules.

Methods

impl<T: Hash + Eq + Clone, S, A> Graph<T, S, A>

pub fn new() -> Self

Creates an empty Graph with no vertices or edges.

pub fn find_node<'s>(&'s self, state: &T) -> Option<Node<'s, T, S, A>>

Gets a node handle for the given game state.

If state does not correspond to a known game state, returns None.

pub fn find_node_mut<'s>(
&'s mut self,
state: &T
) -> Option<MutNode<'s, T, S, A>>

Gets a mutable node handle for the given game state.

If state does not correspond to a known game state, returns None.

pub fn add_node<'s>(&'s mut self, state: T, data: S) -> MutNode<'s, T, S, A>

Adds a vertex (with no parents or children) for the given game state and data and returns a mutable handle for it.

If state is already known, returns a mutable handle to that state, ignoring the data parameter. As a result, this method is guaranteed to return a handle for a root vertex only when state is a novel game state.

pub fn add_edge<'s, F, G>(
&'s mut self,
source: T,
source_data: F,
dest: T,
dest_data: G,
edge_data: A
) -> MutEdge<'s, T, S, A> where
F: for<'b> FnOnce(Node<'b, T, S, A>) -> S,
G: for<'b> FnOnce(Node<'b, T, S, A>) -> S,

Adds an edge from the vertex with state data source to the vertex with state data dest. If vertices are not found for source or dest, they are added, with the data provided by source_data and dest_data callbacks.

The edge that is created will have the data edge_data. Returns a mutable edge handle for that edge.

pub fn vertex_count(&self) -> usize

Returns the number of vertices in the graph.

pub fn edge_count(&self) -> usize

Returns the number of edges in the graph.