Struct search_graph::MutChildList

pub struct MutChildList<'a, T, S, A> where T: Hash + Eq + Clone + 'a, S: 'a, A: 'a {
    // some fields omitted
}

A traversible list of a vertex's outgoing edges.

Methods

impl<'a, T, S, A> MutChildList<'a, T, S, A> where T: Hash + Eq + Clone + 'a, S: 'a, A: 'a

fn len(&self) -> usize

Returns the number of outgoing eges.

fn is_empty(&self) -> bool

Returns true iff there are no outgoing edges.

fn get_edge<'s>(&'s self, i: usize) -> Edge<'s, T, S, A>

Returns an edge handle for the ith edge.

fn get_edge_mut<'s>(&'s mut self, i: usize) -> MutEdge<'s, T, S, A>

Returns an edge handle for the ith edge. Its lifetime will be limited to a local borrow of self.

fn to_edge(self, i: usize) -> MutEdge<'a, T, S, A>

Returns an edge handle for the ith self is consumed, and the return value's lifetime will be the same as that of self.

fn get_source_node<'s>(&'s self) -> Node<'s, T, S, A>

Returns a node handle for the vertex these edges originate from. Its lifetime will be limited to a local borrow of self.

fn get_source_node_mut<'s>(&'s mut self) -> MutNode<'s, T, S, A>

Returns a mutable node handle for the vertex these edges originate from. Its lifetime will be limited to a local borrow of self.

fn to_source_node(self) -> MutNode<'a, T, S, A>

Returns a mutable node handle for the vertex these edges originate from. self is consumed, and the return value's lifetime will be the same as that of self.

fn iter<'s>(&'s self) -> ChildListIter<'s, T, S, A>

Returns an iterator over child edges.

fn add_child<'s>(&'s mut self, data: A) -> MutEdge<'s, T, S, A>

Appends an unexpanded edge to this list of children and returns a mutable handle to it. Its lifetime will be limited to a local borrow of self.

An unexpanded edge is one with known source but unknown target. Unexpanded edges may be expanded by resolving their target (which will be a Target::Unexpanded(EdgeExpander)) and expanding the edge.

fn to_add(self, data: A) -> MutEdge<'a, T, S, A>

Appends unexpanded an edge to the vertex's' children and returns a mutable handle to it. self will be consumed, and the return value's lifetime will be equal to that of self.

An unexpanded edge is one with known source but unknown target. Unexpanded edges may be expanded by resolving their target (which will be a Target::Unexpanded(EdgeExpander)) and expanding the edge.