Struct search_graph::MutEdge

pub struct MutEdge<'a, T, S, A> where T: Hash + Eq + Clone + 'a, S: 'a, A: 'a {
    // some fields omitted
}

Mutable handle to a graph edge ("edge handle").

This zipper-like type enables traversal of a graph along the edge's source and target vertices.

It enables local graph mutation, whether via mutation of edge data or mutation of graph topology (adding vertices). Vertices may be added to unexpanded edges using the handle returned by get_target_mut or to_target.

Methods

impl<'a, T, S, A> MutEdge<'a, T, S, A> where T: Hash + Eq + Clone + 'a, S: 'a, A: 'a

fn get_id(&self) -> usize

Returns an immutable ID that is guaranteed to identify this vertex uniquely within its graph. This ID may change when the graph is mutated.

fn get_data(&self) -> &A

Returns the data at this edge.

fn get_data_mut(&mut self) -> &mut A

Returns the data at this edge, mutably.

fn get_target<'s>(&'s self) -> Target<Node<'s, T, S, A>, ()>

Returns the target of this edge. If the edge is unexpanded, no data will be available. If it is expanded, a node handle will be available, with its lifetime limited to a local borrow of self.

fn get_target_mut<'s>(&'s mut self) -> Target<MutNode<'s, T, S, A>, EdgeExpander<'s, T, S, A>>

Returns the target of this edge. If the edge is unexpanded, an EdgeExpander will be provided. If it is expanded, a mutable node handle will be available. In both cases, lifetimes will be limited to a local borrow of self.

fn to_target(self) -> Target<MutNode<'a, T, S, A>, EdgeExpander<'a, T, S, A>>

Returns the target of this edge. If the edge is unexpanded, an EdgeExpander will be provided. If it is expanded, a mutable node handle will be available. In both cases self is consumed, and the return value's lifetime will be the same as that of self.

fn get_source<'s>(&'s self) -> Node<'s, T, S, A>

Returns a node handle for the source of this edge. Its lifetime will be limited to a local borrow of self.

fn get_source_mut<'s>(&'s mut self) -> MutNode<'s, T, S, A>

Returns a mutable node handle for the source of this edge. Its lifetime will be limited to a local borrow of self.

fn to_source(self) -> MutNode<'a, T, S, A>

Returns a mutable node handle for the source of this edge. self is consumed, and the return value's lifetime will be equal to that of self.

fn to_edge(self) -> Edge<'a, T, S, A>

Returns a non-mutating edge obtained by converting this edge. self is consumed, and the return value's lifetime will be the same as that of self. The source graph is still considered to have a mutable borrow in play, but the resulting edge can be cloned freely.