Module search_graph::view

Provides an editable view of both graph topology and data, while simultaneously allowing multiple live references into the graph.

References into the graph are provided by NodeRef and EdgeRef. They are created with respect to a View, which wraps around a mutable borrow of a Graph. They may only be dereferenced with respect to the view that created them, and operations on a View that would invalidate these references consume the View.

Basic usage

To create a View, use one of the functions defined in this module: of_graph, of_node, or of_edge.

To use a NodeRef or EdgeRef with a View, pass the reference to an appropriate function on the View or index into the View directly:

let mut graph: Graph<u32, String, String> = Graph::new();
view::of_graph(&mut graph, |mut view| {
  let root: view::NodeRef<'_> = view.append_node(0, "root_data".into());
  assert_eq!(view[root], "root_data");
  let child: view::NodeRef<'_> = view.append_node(10, "child_data".into());
  assert_eq!(view.node_data(child), "child_data");
  let edge: view::EdgeRef<'_> = view.append_edge(root, child, "edge_data".into());
  assert_eq!(view[edge], "edge_data");
});

Relationship with search_graph::mutators

The mutators module provides another read-write interface to a Graph, in which cursors into the graph directly own a mutable reference to it. Because these cursors own a mutable reference to a common underlying Graph, it is difficult to have more than one cursor active at a time while still satisfying the borrow checker. This makes it tricky to retain multiple cursors into a graph if even one of them allows the graph to be mutated.

The cursors in mutators may be converted into a View by the of_node and of_edge functions. For example:

let mut graph: Graph<u32, String, String> = Graph::new();
let root: MutNode<'_, u32, String, String> =
  graph.add_node(0, "root_data".into());
view::of_node(root, |view, node| {
  assert_eq!(view[node], "root_data");
});

A View may be transformed into a cursor from mutators by calling into_node, into_edge, into_append_node or into_append_edge. These methods consume a View and release its borrow of a Graph back to a stand-alone cursor type. For example:

let mut graph: Graph<u32, String, String> = Graph::new();
let mut node: MutNode<'_, u32, String, String> =
  graph.add_node(0, "root_data".into());
node = view::of_node(node, |view, node| {
  assert_eq!(view[node], "root_data");
  view.into_append_node(100, "child_data".into())
});
assert_eq!(node.get_data(), "child_data");

Structs

EdgeRef
Reference to a graph edge that is licensed by a View. Only the View that an EdgeRef is associated with can dereference that EdgeRef.

NodeRef
Reference to a graph vertex that is licensed by a View.

View
An editable view of a graph.

Functions

of_edge
Applies a function over a MutEdge and a view of its containing graph and returns the function's result.

of_graph
Applies a function over a view of Graph and returns its result.

of_node
Applies a function over a MutNode and a view of its containing graph and returns the function's result.