Struct search_graph::view::View

pub struct View<'a, 'id, T: Hash + Eq + Clone, S, A> where
    'a: 'id,  { /* fields omitted */ }

An editable view of a graph.

A View wraps around a mutable borrow of a Graph and enables taking references into the graph and performing mutation operations on it. Mutations that invalidate references into the graph consume the view entirely. Generally speaking, it is possible to grow the graph (add nodes or edges) without invalidating references into it. See this module's documentation for more details about how to create and use a View.

Methods

impl<'a, 'id, T: Hash + Eq + Clone, S, A> View<'a, 'id, T, S, A> where
'a: 'id,

pub fn find_node(&self, state: &T) -> Option<NodeRef<'id>>

Returns a reference to the node for the given game state that is already in the graph, or None if there is no such node.

pub fn find_edge(
&self,
source: NodeRef<'id>,
target: NodeRef<'id>
) -> Option<EdgeRef<'id>>

Returns a reference to an edge between the given nodes that is already in the graph, or None if there is no such edge.

pub fn append_node(&mut self, state: T, data: S) -> NodeRef<'id>

Adds a node for the given game state with the given data, returning a reference to the node after it is added. If such a node already exists, no node is added to the graph, and a reference to the existing node is returned.

pub fn into_node(self, node: NodeRef<'id>) -> MutNode<'a, T, S, A>

Consumes this view and returns a MutNode.

pub fn into_append_node(self, state: T, data: S) -> MutNode<'a, T, S, A>

Consumes this view and adds a node as if append_node had been called. Returns a MutNode that points to the node that is created.

pub fn into_edge(self, edge: EdgeRef<'id>) -> MutEdge<'a, T, S, A>

Consumes this view and returns a MutEdge.

pub fn append_edge(
&mut self,
source: NodeRef<'id>,
target: NodeRef<'id>,
edge_data: A
) -> EdgeRef<'id>

Adds an edge between the given nodes, returning a reference to it after it is added.

pub fn into_append_edge(
self,
source: NodeRef<'id>,
target: NodeRef<'id>,
edge_data: A
) -> MutEdge<'a, T, S, A>

Consumes this view and adds an edge as if append_edge had been called. Returns a MutEdge that points to the edge that is created.

pub fn node_state(&self, node: NodeRef<'id>) -> &T

Returns a reference to the game state that node is associated with.

pub fn node_data(&self, node: NodeRef<'id>) -> &S

Returns a reference to the data (usually statistics or payout information) for node.

pub fn node_data_mut(&mut self, node: NodeRef<'id>) -> &mut S

Returns a mutable reference to the data (usually statistics or payout information) for node.

pub fn edge_data(&self, edge: EdgeRef<'id>) -> &A

Returns a reference to the data (usually statistics or payout information) for edge.

pub fn edge_data_mut(&mut self, edge: EdgeRef<'id>) -> &mut A

Returns a mutable reference to the data (usually statistics or payout information) for edge.

pub fn edge_source(&self, edge: EdgeRef<'id>) -> NodeRef<'id>

Returns a reference to the node that edge originates from.

pub fn edge_target(&self, edge: EdgeRef<'id>) -> NodeRef<'id>

Returns a reference to the node that edge terminates on.

pub fn child_count(&self, node: NodeRef<'id>) -> usize

Returns the number of children (outgoing edges) that node has.

pub fn children<'s>(
&'s self,
node: NodeRef<'id>
) -> impl Iterator<Item = EdgeRef<'id>> + 's

Returns an iterator over the children (outgoing edges) that node has.

let mut g: Graph<String, String, String> = Graph::new();
view::of_graph(&mut g, |mut v| {
  let root = v.append_node("root_state".into(), "root_data".into());
  let child1 = v.append_node("child1_state".into(), "child1_data".into());
  let child2 = v.append_node("child2_state".into(), "child2_data".into());
  let child3 = v.append_node("child3_state".into(), "child3_data".into());
  v.append_edge(root, child1, "edge1_data".into());
  v.append_edge(root, child2, "edge2_data".into());
  v.append_edge(root, child3, "edge3_data".into());
  let edge_data: Vec<&String> = v.children(root).map(|e| v.edge_data(e)).collect();
  assert_eq!(edge_data, vec!["edge1_data", "edge2_data", "edge3_data"]);
  let child_data: Vec<&String> =
    v.children(root).map(|e| v.node_data(v.edge_target(e))).collect();
  assert_eq!(child_data, vec!["child1_data", "child2_data", "child3_data"]);
});

pub fn parent_count(&self, node: NodeRef<'id>) -> usize

Returns the number of parents (incoming edges) that node has.

pub fn parents<'s>(
&'s self,
node: NodeRef<'id>
) -> impl Iterator<Item = EdgeRef<'id>> + 's

Returns an iterator over the parents (incoming edges) that node has.

let mut g: Graph<String, String, String> = Graph::new();
view::of_graph(&mut g, |mut v| {
  let child = v.append_node("child_state".into(), "child_data".into());
  let parent1 = v.append_node("parent1_state".into(), "parent1_data".into());
  let parent2 = v.append_node("parent2_state".into(), "parent2_data".into());
  let parent3 = v.append_node("parent3_state".into(), "parent3_data".into());
  v.append_edge(parent1, child, "edge1_data".into());
  v.append_edge(parent2, child, "edge2_data".into());
  v.append_edge(parent3, child, "edge3_data".into());
  let edge_data: Vec<&String> = v.parents(child).map(|e| v.edge_data(e)).collect();
  assert_eq!(edge_data, vec!["edge1_data", "edge2_data", "edge3_data"]);
  let parent_data: Vec<&String> =
    v.parents(child).map(|e| v.node_data(v.edge_source(e))).collect();
  assert_eq!(parent_data, vec!["parent1_data", "parent2_data", "parent3_data"]);
});

pub fn retain_reachable_from<I: IntoIterator<Item = NodeRef<'id>>>(
self,
roots: I
)

Deletes all graph components that are not reachable by a traversal starting from each of roots.