Crate search_graph

This crate provides support for search (in the sense of AI or optimization) over a space of discrete, enumerated states. Its main data structure is Graph, a directed graph with content-addressable vertices.

Several efficient, type-safe interfaces are provided in submodules:

view provides a type- and memory-safe view of a Graph.

nav provides a read-only cursor-based view of a Graph. (If you have seen zippers in other contexts, this pattern should be familiar).

mutators is a read-write analogue of nav.

Modules

mutators
Support for navigation of a graph that allows modifications to graph topology and full read-write access to graph data.

nav
Support for navigation of a graph without allowing modifications to graph topology.

search
Data structures and algorithms for local graph search.

view
Provides an editable view of both graph topology and data, while simultaneously allowing multiple live references into the graph.

Structs

Graph
A directed graph over a space of discrete, enumerated states.