Struct search_graph::mutators::MutChildList

pub struct MutChildList<'a, T: Hash + Eq + Clone + 'a, S: 'a, A: 'a> { /* fields omitted */ }

A traversible list of a vertex's outgoing edges.

Methods

impl<'a, T: Hash + Eq + Clone + 'a, S: 'a, A: 'a> MutChildList<'a, T, S, A>

pub fn len(&self) -> usize

Returns the number of outgoing eges.

pub fn is_empty(&self) -> bool

Returns true iff there are no outgoing edges.

pub fn get_edge<'s>(&'s self, i: usize) -> Edge<'s, T, S, A>

Returns an edge handle for the ith edge.

pub fn get_edge_mut<'s>(&'s mut self, i: usize) -> MutEdge<'s, T, S, A>

Returns an edge handle for the ith edge. Its lifetime will be limited to a local borrow of self.

pub fn to_edge(self, i: usize) -> MutEdge<'a, T, S, A>

Returns an edge handle for the ith self is consumed, and the return value's lifetime will be the same as that of self.

pub fn get_source_node<'s>(&'s self) -> Node<'s, T, S, A>

Returns a node handle for the vertex these edges originate from. Its lifetime will be limited to a local borrow of self.

pub fn get_source_node_mut<'s>(&'s mut self) -> MutNode<'s, T, S, A>

Returns a mutable node handle for the vertex these edges originate from. Its lifetime will be limited to a local borrow of self.

pub fn to_source_node(self) -> MutNode<'a, T, S, A>

Returns a mutable node handle for the vertex these edges originate from. self is consumed, and the return value's lifetime will be the same as that of self.

Important traits for ChildListIter<'a, T, S, A>
impl<'a, T, S, A> Iterator for ChildListIter<'a, T, S, A> where
T: Hash + Eq + Clone + 'a,
S: 'a,
A: 'a, type Item = Edge<'a, T, S, A>;

pub fn iter<'s>(&'s self) -> ChildListIter<'s, T, S, A>

Returns an iterator over child edges.

pub fn add_child<'s, F>(
&'s mut self,
child_label: T,
f: F,
edge_data: A
) -> MutEdge<'s, T, S, A> where
F: FnOnce() -> S,

Adds a child edge to the vertex labeled by child_label. If no such vertex exists, it is created and associated with the data returned by f. Returns a mutable edge handle for the new edge, with a lifetime limited to a borrow of self.

pub fn to_add_child<F>(
self,
child_label: T,
f: F,
edge_data: A
) -> MutEdge<'a, T, S, A> where
F: FnOnce() -> S,

Adds a child edge to the vertex labeled by child_label. If no such vertex exists, it is created and associated with the data returned by f. Returns a mutable edge handle for the new edge.