Struct search_graph::search::Stack

pub struct Stack<'a, T: 'a + Hash + Eq + Clone, S: 'a, A: 'a> { /* fields omitted */ }

Tracks the path through a graph that is followed when performing local search.

In this case, "local search" is a process that starts focused on a single vertex and incrementally updates which vertex is the focus by traversing parent or child edges. The history of such operations can be described as a series of (vertex, edge) pairs, and a Stack encapsulates this history.

A Stack points to a graph vertex, whose incidental edges can then be traversed. Operations which modify graph topology may cause the search path's internal state to fall out of sync with the graph's state, so graph elements are exposed using the read-only Node and Edge types.

A path may be consumed to yield a read-write view of the underlying graph with the to_head method.

Methods

impl<'a, T: 'a + Hash + Eq + Clone, S: 'a, A: 'a> Stack<'a, T, S, A>

pub fn new(node: MutNode<'a, T, S, A>) -> Self

Creates a new Stack from a mutable reference into a graph.

pub fn len(&self) -> usize

Returns the number of elements in the path. Since a path always has a head, there is always at least 1 element.

pub fn pop<'s>(&'s mut self) -> Option<Edge<'s, T, S, A>>

Removes the most recently traversed element from the path, if any. Returns a handle for any edge that was removed.

pub fn head<'s>(&'s self) -> Node<'s, T, S, A>

Returns a read-only view of the head element.

pub fn to_head(self) -> MutNode<'a, T, S, A>

Consumes the path and returns a mutable view of its head.

pub fn push<'s, F, E>(
&'s mut self,
f: F
) -> Result<Option<Edge<'s, T, S, A>>, SearchError<E>> where
F: FnMut(&Node<'s, T, S, A>) -> Result<Option<Traversal>, E>,
E: Error,

Grows the path by consulting a function of the current head. If this function f returns Ok(Some(Traversal::Child(i))), then the ith child of the current head is pushed onto the path. If it returns Ok(Some(Traversal::Parent(i))), then the ith parent of the current head is pushed onto the path.

The decision not to traverse any edge may be made by returning Ok(None), while Err(E) should be returned for any errors.

Returns an Ok(Option(e)) for any edge e that is traversed, or Err(e) if an error was encountered.

Important traits for StackIter<'a, 's, T, S, A>
impl<'a, 's, T, S, A> Iterator for StackIter<'a, 's, T, S, A> where
T: 'a + Hash + Eq + Clone,
S: 'a,
A: 'a,
'a: 's, type Item = StackItem<'s, T, S, A>;

pub fn iter<'s>(&'s self) -> StackIter<'a, 's, T, S, A>

Returns an iterator over path elements. Iteration is in order of traversal (i.e., the last element of the iteration is the path head).

pub fn item<'s>(&'s self, i: usize) -> Option<StackItem<'s, T, S, A>>

Returns the ith item of the path. Path items are indexed in order of traversal (i.e., the last element is the path head).