Struct search_graph::Graph

pub struct Graph<T, S, A> where T: Hash + Eq + Clone {
    // some fields omitted
}

A search graph.

Supports incremental rollout of game state topology and vertex de-duplication with transposition tables. Limited support is provided for deletion of components and compaction in memory after deletion.

T: The type of game states. It is required to derive Hash and Eq to so that it may be stored in a hashtable, where game states are looked up to support de-duplication of game states. It is required to derive Clone to accommodate a limitation of the HashMap interface.

S: The type of graph vertex data.

A: The type of graph edge data.

Vertices are addressed by content. To examine graph contents, obtain a node handle with get_node. To modify graph contents, add new root vertices with add_root and retrieve extant vertices with get_node_mut.

Methods

impl<T, S, A> Graph<T, S, A> where T: Hash + Eq + Clone

fn new() -> Self

Creates an empty Graph with no vertices or edges.

fn get_node<'s>(&'s self, state: &T) -> Option<Node<'s, T, S, A>>

Gets a node handle for the given game state.

If state does not correspond to a known game state, returns None.

fn get_node_mut<'s>(&'s mut self, state: &T) -> Option<MutNode<'s, T, S, A>>

Gets a mutable node handle for the given game state.

If state does not correspond to a known game state, returns None.

fn add_root<'s>(&'s mut self, state: T, data: S) -> MutNode<'s, T, S, A>

Adds a root vertex (one with no parents) for the given game state and data and returns a mutable handle for it.

If state is already known, returns a mutable handle to that state, ignoring the data parameter. As a result, this method is guaranteed to return a handle for a root vertex only when state is a novel game state.

fn add_edge<'s, F, G>(&'s mut self, source: T, source_data: F, dest: T, dest_data: G, edge_data: A) -> MutEdge<'s, T, S, A> where F: for<'b> FnOnce(Node<'b, T, S, A>) -> S, G: for<'b> FnOnce(Node<'b, T, S, A>) -> S

Adds an edge from the vertex with state data source to the vertex with state data dest. If vertices are not found for source or dest, they are added, with the data provided by source_data and dest_data callbacks.

The edge that is created will have the data edge_data. Returns a mutable edge handle for that edge.

fn vertex_count(&self) -> usize

Returns the number of vertices in the graph.

fn edge_count(&self) -> usize

Returns the number of edges in the graph.

fn retain_reachable_from(&mut self, roots: &[T])

Deletes all graph components that are not reachable by traversal starting from each vertex corresponding to the game states in roots.

Game states in roots which do not have a corresponding vertex are ignored.